+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version 26.6.12.1
 
-## Configured Adguard Home and Tailscale
+## Added Security updates to the System
+- Created a new file called `security.nix` in `system/modules/`
+- Added Kernel Hardening
+- Added Fail2Ban
+- Added Memory Security
 
-They both are used in connecting with the new domain, `.home`. This domain is accessible through the tailscale app.